// src/app/modules/UserAndOrderManagement/UserAndOrderManagement.interface.ts
export type FullName = {
  firstName: string;
  lastName: string;
}

export type Address = {
  street: string;
  city: string;
  country: string;
}

export type Orders = {
  productName: string;
  price: number;
  quantity: number;
}

export type UserAndOrder = {
  userId: number;
  username: string;
  password: string;
  fullName: FullName;
  age: number;
  email: string;
  isActive: boolean;
  hobbies: string[];
  address: Address;
  orders: Orders[];
}